requirement.txt
Django=3.0.8
PyJWT=1.7.1
asgiref=3.2.10
djangorestframework=3.11.0
djangorestframework-jwt=1.11.0
pip=20.1.1
pytz=2020.1
setuptools=49.2.0
sqlparse=0.3.1
django-user-activity-log2